# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0

paths:
  /products:
    post:
      summary: Создание товара
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      summary: Список товаров
      operationId: getProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /products/{id}:
    get:
      summary: Получение товара по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Обновление товара
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Мягкое удаление товара
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Товар успешно удален
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true

    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    ProductCreate:
      type: object
      required: [name, price, stock, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
          default: ACTIVE

    ProductUpdate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PaginatedProductResponse:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
        page:
          type: integer
        size:
          type: integer